+++ src/lib/Icons/IconRefresh.stories.svelte
@@ -1,0 +1,22 @@
+<script context="module">
+  import IconRefresh from "./IconRefresh.svelte";
+
+  export const meta = {
+    title: "Icons/Icons/IconRefresh",
+    component: IconRefresh,
+    tags: ["!dev", "!autodocs"],
+    parameters: {
+      githubLink: {
+        url: "/Icons/IconRefresh.svelte"
+      }
+    }
+  };
+</script>
+
+<script>
+  import { Story } from "@storybook/addon-svelte-csf";
+</script>
+
+<Story name="Default" let:args>
+  <IconRefresh {...args} />
+</Story>